+++ lib/helpers.js
@@ -56,21 +56,6 @@ function version () {
   }
 }
 
-function filterForNonReserved (reserved, options) {
-  // Filter out properties that are not reserved.
-  // Reserved values are passed in at call site.
-
-  var object = Object.create(null)
-  for (var i in options) {
-    var notReserved = (reserved.indexOf(i) === -1)
-    if (notReserved) {
-      object[i] = options[i]
-    }
-  }
-  return object
-}
-
-exports.filterForNonReserved = filterForNonReserved
 exports.paramsHaveRequestBody = paramsHaveRequestBody
 exports.safeStringify = safeStringify
 exports.md5 = md5
